Impact-Dashboard-Backend/pyproject.toml
[tool.poetry]
name = "impact-dashboard-backend"
version = "0.1.0"
description = "The backend pipleline of the impact dashboard for Coldtivate"
authors = ["divinefavourodion <divineod9@gmail.com>"]
readme = "README.md"
packages = [{include = "impact_dashboard_backend"}]

[tool.poetry.dependencies]
python = "^3.8"
appdirs = "1.4.4"
autopep8 = "2.0.0"
certifi = "2022.9.24"
charset-normalizer = "2.1.1"
idna = "3.4"
lxml = "4.9.1"
multitasking = "0.0.11"
numpy = "1.23.4"
pandas = "1.5.1"
psycopg2-binary = "2.9.4"
pycodestyle = "2.9.1"
python-dateutil = "2.8.2"
python-dotenv = "^1.0.0"
pytz = "2022.5"
requests = "2.28.1"
six = "1.16.0"
tomli = "2.0.1"
urllib3 = "1.26.12"
yfinance = "0.1.85"


[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
pylint = "^2.17.4"
autoflake = "^2.1.1"
pyupgrade = "^3.4.0"
isort = "^5.12.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
